[3.0.0-SNAPSHOT] Micronaut updated to 3.0.0 LegacyArangoClusterHealthIndicator.java removed Micronaut 3.0.0 integration fixes All timeout related properties changed to Duration
