+++ gradle.properties
@@ -1,6 +1,6 @@
 groupId=com.github.goodforgod
 artifactId=micronaut-arangodb
-artifactVersion=2.4.0
+artifactVersion=3.0.0-SNAPSHOT
 
 
 ##### GRADLE
