+++ src/main/java/io/micronaut/configuration/arango/ArangoAccessorAsyncFactory.java
@@ -45,7 +45,7 @@ public class ArangoAccessorAsyncFactory {
 
             return builder.build();
         } catch (IOException e) {
-            throw new ConfigurationException(e.getMessage());
+            throw new ConfigurationException(e.getMessage(), e);
         }
     }
 }
