+++ src/main/java/io/micronaut/configuration/arango/ArangoAsyncConfiguration.java
@@ -3,8 +3,8 @@ package io.micronaut.configuration.arango;
 import io.micronaut.configuration.arango.ssl.ArangoSSLConfiguration;
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Requires;
+import jakarta.inject.Inject;
 
-import javax.inject.Inject;
 
 /**
  * ArangoDB Async configuration class.
