+++ src/main/java/io/micronaut/configuration/arango/ArangoConfiguration.java
@@ -4,8 +4,8 @@ import com.arangodb.Protocol;
 import io.micronaut.configuration.arango.ssl.ArangoSSLConfiguration;
 import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.context.annotation.Requires;
+import jakarta.inject.Inject;
 
-import javax.inject.Inject;
 import java.util.Properties;
 
 /**
