+++ src/main/java/io/micronaut/configuration/arango/ArangoDatabaseInitializer.java
@@ -9,7 +9,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.PostConstruct;
+import java.time.Duration;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
@@ -53,7 +55,7 @@ public class ArangoDatabaseInitializer {
     protected void initializeDatabaseSynchronously(ArangoClientAsync clientAsync,
                                                    ArangoAsyncConfiguration configuration) {
         final String database = configuration.getDatabase();
-        final int timeout = configuration.getCreateDatabaseTimeoutInMillis();
+        final Duration timeout = configuration.getCreateDatabaseTimeout();
 
         try {
             logger.debug("Arango Database '{}' initialization starting...", database);
@@ -62,7 +64,7 @@ public class ArangoDatabaseInitializer {
                     .thenCompose(exist -> exist
                             ? CompletableFuture.completedFuture(true)
                             : clientAsync.db().create())
-                    .get(timeout, TimeUnit.MILLISECONDS);
+                    .get(timeout.toMillis(), TimeUnit.MILLISECONDS);
             final long tookTime = System.currentTimeMillis() - startTime;
             logger.debug("Arango Database '{}' creation took '{}' millis", database, tookTime);
         } catch (TimeoutException e) {
@@ -70,7 +72,8 @@ public class ArangoDatabaseInitializer {
         } catch (ArangoDBException e) {
             throw e;
         } catch (Exception e) {
-            throw new ApplicationStartupException("Arango Database initialization failed due to: " + e.getMessage(), e);
+            final Throwable cause = (e instanceof CompletionException) ? e.getCause() : e;
+            throw new ApplicationStartupException("Arango Database initialization failed due to: " + cause.getMessage(), cause);
         }
     }
 }
